Fix #1105, Fix action progress data error issue

+++ smart-metastore/src/main/java/org/smartdata/metastore/dao/ActionDao.java
@@ -235,7 +235,7 @@ public class ActionDao {
     parameters.put("create_time", actionInfo.getCreateTime());
     parameters.put("finished", actionInfo.isFinished());
     parameters.put("finish_time", actionInfo.getFinishTime());
-    parameters.put("progress", (int) (actionInfo.getProgress()));
+    parameters.put("progress", actionInfo.getProgress());
     return parameters;
   }
 
@@ -256,7 +256,7 @@ public class ActionDao {
       actionInfo.setCreateTime(resultSet.getLong("create_time"));
       actionInfo.setFinished(resultSet.getBoolean("finished"));
       actionInfo.setFinishTime(resultSet.getLong("finish_time"));
-      actionInfo.setProgress(resultSet.getInt("progress"));
+      actionInfo.setProgress(resultSet.getFloat("progress"));
       return actionInfo;
     }
   }
